requirements.txt
APScheduler==3.6.0
pyserial==3.4

